Implemented implicit backface culling of triangle meshes.
This implementation utilises the order of the vertices of each triangle to determine whether the side facing the direction of the ray is the front or back side. The front side is defined as the side where all vertices of the triangle are defined in clockwise order.
The implementation makes use of one of the two integer padding fields of the ray data structure to declare whether the ray needs to cull backfaces or not. Thus to enable backface culling for a ray, the new member variable "int doBackfaceCulling;" needs to be set to true/1 either manually or by using the new method "void SetDoBackfaceCulling(bool _bDoBackfaceCulling)".
If backface culling is disabled via the CMake config switch (RR_ENABLE_BACKFACE_CULL; off by default) all backface culling checks in the intersection kernels are disabled at compile time to make sure that there is no performance impact when not using the feature.